requirements.txt
# Core Django Packages
Django==4.2.25
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-filter==23.2
dj-database-url==2.0.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
gunicorn==21.0.0
whitenoise==6.5.0

# Data Processing
pandas==2.0.3
openpyxl==3.1.2
numpy==1.24.3

# Report Generation
reportlab==4.0.4
xlsxwriter==3.1.2

# Timezone and Scheduling
pytz==2023.3
python-dateutil==2.8.2

# Email and Notifications
celery==5.3.4
redis==5.0.1

# AI & Vector Search (Optional - can be installed separately if needed)
# pgvector==0.2.4
# openai==0.27.8
# httpx==0.24.1
# numpy==1.24.4
